backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
numpy==1.26.4
rasterio==1.3.10
pyproj==3.6.1
scipy>=1.11.0
PyWavelets>=1.5.0
numba>=0.58.0
joblib>=1.3.0
PyJWT>=2.8.0
Werkzeug>=3.0.0
Pillow>=10.0.0
